gst-python: Autospec creation for update from version 1.18.4 to version 1.18.5
The latest bug-fix release in the 1.18 series is 1.18.5 and was released
on 8 September 2021.

See https://gstreamer.freedesktop.org/releases/1.18/ for the latest
version of this document.

Last updated: Wednesday 8 September 2021, 11:00 UTC (log)

Introduction

The GStreamer team is proud to announce a new major feature release in
the stable 1.x API series of your favourite cross-platform multimedia
framework!

As always, this release is again packed with many new features, bug

(NEWS truncated at 15 lines)

=== release 1.18.5 ===

2021-09-08 20:05:12 +0100  Tim-Philipp Müller <tim@centricular.com>

	* ChangeLog:
	* NEWS:
	* RELEASE:
	* gst-python.doap:
	* meson.build:
	  Release 1.18.5

2021-03-15 19:49:35 +0000  Tim-Philipp Müller <tim@centricular.com>

	* meson.build:
	  Back to development